Fix DOWNLOAD_ switches in modules/common/deps.sh

+++ modules/common/deps.sh
@@ -5,11 +5,6 @@
 . ../../include/common.sh
 . .envrc
 
-if [[ "$DOWNLOAD_BINS" == "false" ]]; then
-    ok "Skipping downloading bins, using host binaries"
-    exit 0
-fi
-
 if [[ "$OSTYPE" == "darwin"* ]]; then
     export HELM_OS_TYPE="${HELM_OS_TYPE:-darwin-amd64}"
     export KUBECTL_OS_TYPE="${KUBECTL_OS_TYPE:-darwin}"
@@ -24,50 +19,56 @@ else
     export YQ_OS_TYPE="${YQ_OS_TYPE:-linux_amd64}"
 fi
 
-if [ ! -e "bin/helm" ]; then
-    curl -L https://get.helm.sh/helm-${HELM_VERSION}-${HELM_OS_TYPE}.tar.gz | tar zxf -
-    mv $HELM_OS_TYPE/helm bin/
-    mv $HELM_OS_TYPE/tiller bin/
-    rm -rf "$HELM_OS_TYPE"
-    helm version --client
+
+if [[ "$DOWNLOAD_BINS" == "false" ]]; then
+    ok "Skipping downloading bins, using host binaries"
+else
+    if [ ! -e "bin/helm" ]; then
+        curl -L https://get.helm.sh/helm-${HELM_VERSION}-${HELM_OS_TYPE}.tar.gz | tar zxf -
+        mv $HELM_OS_TYPE/helm bin/
+        mv $HELM_OS_TYPE/tiller bin/
+        rm -rf "$HELM_OS_TYPE"
+        helm version --client
+    fi
+
+    if [ ! -e "bin/kubectl" ]; then
+        curl -LO https://storage.googleapis.com/kubernetes-release/release/${KUBECTL_VERSION}/bin/${KUBECTL_OS_TYPE}/amd64/kubectl
+        mv kubectl bin/
+        chmod +x bin/kubectl
+        kubectl version 2>&1 | grep Client || true
+    fi
+
+    if [ ! -e "bin/cf" ]; then
+        curl -L "https://packages.cloudfoundry.org/stable?release=${CFCLI_OS_TYPE}-binary&source=github" | tar -zx
+        mv cf bin/
+        rm -rf "$CFCLI_OS_TYPE" LICENSE NOTICE
+        chmod +x bin/cf
+        cf version
+    fi
 fi
 
-if [ ! -e "bin/kubectl" ]; then
-    curl -LO https://storage.googleapis.com/kubernetes-release/release/${KUBECTL_VERSION}/bin/${KUBECTL_OS_TYPE}/amd64/kubectl
-    mv kubectl bin/
-    chmod +x bin/kubectl
-    kubectl version 2>&1 | grep Client || true
-fi
-
-if [ ! -e "bin/cf" ]; then
-    curl -L "https://packages.cloudfoundry.org/stable?release=${CFCLI_OS_TYPE}-binary&source=github" | tar -zx
-    mv cf bin/
-    rm -rf "$CFCLI_OS_TYPE" LICENSE NOTICE
-    chmod +x bin/cf
-    cf version
-fi
 
 if [[ "$DOWNLOAD_CATAPULT_DEPS" == "false" ]]; then
     ok "Skipping downloading catapult dependencies, using host binaries"
-    exit 0
+else
+    bazelpath=bin/bazel
+    if [ ! -e "$bazelpath" ]; then
+        wget "https://github.com/bazelbuild/bazel/releases/download/${BAZEL_VERSION}/bazel-${BAZEL_VERSION}-linux-x86_64" -O $bazelpath
+        chmod +x $bazelpath
+    fi
+
+    yamlpatchpath=bin/yaml-patch
+    if [ ! -e "$yamlpatchpath" ]; then
+        wget "https://github.com/krishicks/yaml-patch/releases/download/v0.0.10/yaml_patch_${YAMLPATCH_OS_TYPE}" -O $yamlpatchpath
+        chmod +x $yamlpatchpath
+    fi
+
+    yqpath=bin/yq
+    if [ ! -e "$yqpath" ]; then
+        wget "https://github.com/mikefarah/yq/releases/download/2.4.0/yq_${YQ_OS_TYPE}" -O $yqpath
+        chmod +x $yqpath
+    fi
 fi
 
-bazelpath=bin/bazel
-if [ ! -e "$bazelpath" ]; then
-    wget "https://github.com/bazelbuild/bazel/releases/download/${BAZEL_VERSION}/bazel-${BAZEL_VERSION}-linux-x86_64" -O $bazelpath
-    chmod +x $bazelpath
-fi
-
-yamlpatchpath=bin/yaml-patch
-if [ ! -e "$yamlpatchpath" ]; then
-    wget "https://github.com/krishicks/yaml-patch/releases/download/v0.0.10/yaml_patch_${YAMLPATCH_OS_TYPE}" -O $yamlpatchpath
-    chmod +x $yamlpatchpath
-fi
-
-yqpath=bin/yq
-if [ ! -e "$yqpath" ]; then
-    wget "https://github.com/mikefarah/yq/releases/download/2.4.0/yq_${YQ_OS_TYPE}" -O $yqpath
-    chmod +x $yqpath
-fi
 
 ok "Deps correctly downloaded"
